Move repository into its own module
