+++ mlserver/repository/__init__.py
@@ -1,0 +1,3 @@
+from .repository import ModelRepository, DEFAULT_MODEL_SETTINGS_FILENAME
+
+__all__ = ["ModelRepository", "DEFAULT_MODEL_SETTINGS_FILENAME"]
