+++ tests/repository/test_repository.py
@@ -1,13 +1,11 @@
 import os
 import json
 import pytest
-import shutil
 
 from mlserver.repository import ModelRepository, DEFAULT_MODEL_SETTINGS_FILENAME
 from mlserver.settings import ModelSettings, ENV_PREFIX_MODEL_SETTINGS
 
-from .helpers import get_import_path
-from .conftest import TESTDATA_PATH
+from .utils import get_import_path
 
 
 @pytest.fixture
@@ -40,24 +38,6 @@ def multi_model_folder(model_folder: str, sum_model_settings: ModelSettings) -> 
     return model_folder
 
 
-@pytest.fixture
-def custom_module_folder(model_settings: ModelSettings, tmp_path: str) -> str:
-    # Copy models.py, which acts as custom module
-    src = os.path.join(TESTDATA_PATH, "models.py")
-    dst = os.path.join(tmp_path, "models.py")
-    shutil.copyfile(src, dst)
-
-    # Add modified settings, pointing to local module
-    model_settings_path = os.path.join(tmp_path, DEFAULT_MODEL_SETTINGS_FILENAME)
-    with open(model_settings_path, "w") as f:
-        settings_dict = sum_model_settings.dict()
-        # Point to local module
-        settings_dict["implementation"] = "models.SumModel"
-        f.write(json.dumps(settings_dict))
-
-    return str(tmp_path)
-
-
 async def test_list(
     sum_model_settings: ModelSettings, model_repository: ModelRepository
 ):
@@ -77,19 +57,6 @@ async def test_list(
     assert loaded_model_settings._source == os.path.join(
         model_repository._root, DEFAULT_MODEL_SETTINGS_FILENAME
     )
-
-
-#  async def test_list_custom_module(
-#  custom_module_folder: str, sum_model_settings: ModelSettings
-#  ):
-#  multi_model_repository = ModelRepository(custom_module_folder)
-
-#  settings_list = await multi_model_repository.list()
-
-#  assert len(settings_list) == 1
-#  model_settings = settings_list[0]
-#  assert model_settings.name == sum_model_settings.name
-#  assert get_import_path(model_settings.implementation) == "models.SumModel"
 
 
 async def test_list_multi_model(multi_model_folder: str):
@@ -141,23 +108,6 @@ async def test_list_fallback(
         == sum_model_settings.parameters.version  # type: ignore
     )
     assert default_model_settings._source is None
-
-
-async def test_name_fallback(
-    sum_model_settings: ModelSettings,
-    model_folder: str,
-    model_repository: ModelRepository,
-):
-    # Create empty model-settings.json file
-    model_settings_path = os.path.join(model_folder, DEFAULT_MODEL_SETTINGS_FILENAME)
-    with open(model_settings_path, "w") as model_settings_file:
-        d = sum_model_settings.dict()
-        del d["name"]
-        d["implementation"] = get_import_path(d["implementation"])
-        json.dump(d, model_settings_file)
-
-    model_settings = model_repository._load_model_settings(model_settings_path)
-    assert model_settings.name == os.path.basename(model_folder)
 
 
 async def test_find(
